Move line-break helper into shared utils
The line break helper used to live at src/shared/line-breaks.js; relocate it to src/shared/utils/line-breaks.js so text utilities sit with the rest of the shared utils bundle. Update imports, utils barrel, and shared module docs to match.
