+++ src/parser/tests/parser.test.js
@@ -7,7 +7,7 @@ import { describe, it } from "node:test";
 
 import GMLParser from "../gml-parser.js";
 import GameMakerASTBuilder from "../src/gml-ast-builder.js";
-import { getLineBreakCount } from "../../shared/line-breaks.js";
+import { getLineBreakCount } from "../../shared/utils/line-breaks.js";
 import { getNodeStartIndex } from "../../shared/ast-locations.js";
 
 const currentDirectory = fileURLToPath(new URL(".", import.meta.url));
