+++ src/plugin/src/comments/comment-boundary.js
@@ -7,5 +7,5 @@ export {
     isDocCommentLine,
     isLineComment
 } from "../../../shared/comments.js";
-export { getLineBreakCount } from "../../../shared/line-breaks.js";
+export { getLineBreakCount } from "../../../shared/utils/line-breaks.js";
 export { isObjectLike } from "../../../shared/object-utils.js";
