+++ src/shared/utils/index.js
@@ -1,5 +1,6 @@
 export * from "./array.js";
 export * from "./json.js";
+export * from "./line-breaks.js";
 export * from "./number.js";
 export * from "./numeric-options.js";
 export * from "./object.js";
